backend/face_detection/requirements.txt
Flask
flask-restplus==0.13.0
Flask-Cors==3.0.7
Werkzeug==0.16.0

numpy==1.16.2
opencv-python==4.1.0.25
Pillow==6.0.0
requests==2.21.0
pytest
imutils
dlib
